refactor: split app.go into multiple files based on responsibility
